apidef/importer: new package for blueprint/swagger
These chunks of code are separate from the rest of the main package, so
they make for clean decoupling.

While at it, return errors instead of logging them in most cases.
